zavrsen projekat

+++ composeApp/src/commonMain/kotlin/rs/edu/raf/showtime/domain/quiz/QuizStats.kt
@@ -1,6 +1,6 @@
 package rs.edu.raf.showtime.domain.quiz
 
 data class QuizStats(
-    val bestScore: Int = 0,
+    val bestScore: Double = 0.0,
     val playedCount: Int = 0,
 )
